+++ lib/modules/storage/index.js
@@ -1,62 +1,19 @@
 /* @flow */
 import { NativeModules, NativeEventEmitter } from 'react-native';
 
-import { promisify, noop } from '../utils';
-import { Base, ReferenceBase } from './base';
+import { Base } from './../base';
+import StorageRef from './reference';
+import { promisify, noop } from './../../utils';
 
 const FirestackStorage = NativeModules.FirestackStorage;
 const FirestackStorageEvt = new NativeEventEmitter(FirestackStorage);
 
-class StorageRef extends ReferenceBase {
-  constructor(storage, path) {
-    super(storage.firestack, path);
-
-    this.storage = storage;
-  }
-
-  downloadUrl(): Promise<Object> {
-    const path = this.pathToString();
-    this.log.debug('downloadUrl(', path, ')');
-    return promisify('downloadUrl', FirestackStorage)(this.storage.storageUrl, path)
-      .catch(err => {
-        this.log.error('Error downloading URL for ', path, '.  Error: ', err);
-        throw err;
-      });
-  }
-
-  /**
-   * Downloads a reference to the device
-   * @param {String} downloadPath Where to store the file
-   * @param listener
-   * @return {Promise}
-   */
-  download(downloadPath: string, listener: Function = noop): Promise<Object> {
-    const path = this.pathToString();
-    this.log.debug('download(', path, ') -> ', downloadPath);
-    const listeners = [
-      this.storage._addListener('download_progress', listener),
-      this.storage._addListener('download_paused', listener),
-      this.storage._addListener('download_resumed', listener),
-    ];
-
-    return promisify('downloadFile', FirestackStorage)(this.storage.storageUrl, path, downloadPath)
-      .then((res) => {
-        this.log.debug('res --->', res);
-        listeners.forEach(listener => listener.remove());
-        return res;
-      })
-      .catch(err => {
-        this.log.error('Error downloading ', path, ' to ', downloadPath, '.  Error: ', err);
-        throw err;
-      });
-  }
-}
-
 type StorageOptionsType = {
   storageBucket?: ?string,
 };
+
 export default class Storage extends Base {
-  constructor(firestack: Object, options:StorageOptionsType={}) {
+  constructor(firestack: Object, options: StorageOptionsType = {}) {
     super(firestack, options);
 
     if (this.options.storageBucket) {
@@ -69,14 +26,13 @@ export default class Storage extends Base {
   ref(...path: Array<string>): StorageRef {
     const key = this._pathKey(...path);
     if (!this.refs[key]) {
-      const ref = new StorageRef(this, path);
-      this.refs[key] = ref;
+      this.refs[key] = new StorageRef(this, path);
     }
     return this.refs[key];
   }
 
   /**
-   * Upload a filepath
+   * Upload a file path
    * @param  {string} name     The destination for the file
    * @param  {string} filePath The local path of the file
    * @param  {object} metadata An object containing metadata
@@ -97,7 +53,7 @@ export default class Storage extends Base {
         listeners.forEach(listener => listener.remove());
         return res;
       })
-      .catch(err => {
+      .catch((err) => {
         this.log.error('Error uploading file ', name, ' to ', _filePath, '.  Error: ', err);
         throw err;
       });
@@ -108,8 +64,7 @@ export default class Storage extends Base {
   }
 
   _addListener(evt: string, cb: (evt: Object) => Object): {remove: () => void} {
-    let listener = FirestackStorageEvt.addListener(evt, cb);
-    return listener;
+    return FirestackStorageEvt.addListener(evt, cb);
   }
 
   setStorageUrl(url: string): void {
@@ -134,7 +89,7 @@ export default class Storage extends Base {
   };
 
   get namespace(): string {
-    return 'firestack:storage'
+    return 'firestack:storage';
   }
 }
